refactor: 💡 major profile refactoring and minor cleanup
