+++ src/components/profile/ConsultantPrivateData.tsx
@@ -1,23 +1,20 @@
 import * as React from 'react';
 import { useContext } from 'react';
-import { translate } from '../../resources/scripts/i18n/translate';
 import { UserDataContext } from '../../globalState';
-import './profile.styles';
+import { translate } from '../../resources/scripts/i18n/translate';
+import { Headline } from '../headline/Headline';
+import { Text } from '../text/Text';
 
-export const ProfileDataViewConsultant = () => {
+export const ConsultantPrivateData = () => {
 	const { userData } = useContext(UserDataContext);
 
 	return (
-		<div className="profile__content__item profile__data">
-			<p className="profile__content__title">
-				{translate('profile.data.title')}
-			</p>
-			<div className="profile__data__item">
-				<p className="profile__data__label">
-					{translate('profile.data.userName')}
-				</p>
-				<p className="profile__data__content">{userData.userName}</p>
-			</div>
+		<>
+			<Headline text={'Private Daten'} semanticLevel="5" />
+			<Text
+				text={'Diese Daten können die Ratsuchenden nicht einsehen.'}
+				type="infoLargeAlternative"
+			/>
 			<div className="profile__data__item">
 				<p className="profile__data__label">
 					{translate('profile.data.firstName')}
@@ -66,18 +63,6 @@ export const ProfileDataViewConsultant = () => {
 						: translate('profile.noContent')}
 				</p>
 			</div>
-			<div className="profile__data__item">
-				<p className="profile__data__label">
-					{translate('profile.data.agency')}
-				</p>
-				{userData.agencies.map((item, i) => {
-					return (
-						<p className="profile__data__content" key={i}>
-							{item.name}
-						</p>
-					);
-				})}
-			</div>
-		</div>
+		</>
 	);
 };
